+++ internal/fleet/agent_policy/testdata/TestAccResourceAgentPolicyWithAdvancedMonitoring/remove_advanced_monitoring/main.tf
@@ -1,0 +1,14 @@
+provider "elasticstack" {
+  elasticsearch {}
+  kibana {}
+}
+
+resource "elasticstack_fleet_agent_policy" "test_policy" {
+  name            = var.policy_name
+  namespace       = "default"
+  description     = "Test Agent Policy - No Advanced Monitoring"
+  monitor_logs    = true
+  monitor_metrics = true
+  skip_destroy    = var.skip_destroy
+}
+