+++ internal/fleet/integration_policy/testdata/TestAccResourceIntegrationPolicyWithOutput/update/integration_policy.tf
@@ -86,22 +86,23 @@ resource "elasticstack_fleet_integration_policy" "test_policy" {
   integration_version = elasticstack_fleet_integration.test_policy.version
   output_id           = elasticstack_fleet_output.test_output_updated.output_id
 
-  input {
-    input_id = "tcp-tcp"
-    enabled  = false
-    streams_json = jsonencode({
-      "tcp.generic" : {
-        "enabled" : false
-        "vars" : {
-          "listen_address" : "localhost"
-          "listen_port" : 8085
-          "data_stream.dataset" : "tcp.generic"
-          "tags" : []
-          "syslog_options" : "field: message"
-          "ssl" : ""
-          "custom" : ""
+  inputs = {
+    "tcp-tcp" = {
+      enabled = true
+      streams = {
+        "tcp.generic" = {
+          enabled = true
+          vars = jsonencode({
+            "listen_address" : "localhost"
+            "listen_port" : 8080
+            "data_stream.dataset" : "tcp.generic"
+            "tags" : []
+            "syslog_options" : "field: message"
+            "ssl" : ""
+            "custom" : ""
+          })
         }
       }
-    })
+    }
   }
 }